src/requirements.txt
auto_mix_prep==0.2.0
flwr==0.17.0
matplotlib
mmcv==1.3.16
nibabel==3.2.1
numpy
openpyxl==3.0.9
pandas
Pillow==8.4.0
PyYAML==6.0
scikit_learn
scipy
SimpleITK
scikit-image
torch==1.10.0
torchio==0.18.61
torchvision==0.11.1
tqdm==4.62.3
